requirements.txt.txt
python-telegram-bot==20.7
transformers==4.41.0
torch==2.3.0
accelerate==0.30.0
sentencepiece==0.1.99